Move BlendMode and AlphaCompositing into Colour.Compositing namespace
