--- Assets/Scripts/Colour/AlphaCompositing.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ab19f5d15d2e1f84cb4e329d08ea954a
