--- Assets/Scripts/Colour/BlendMode.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b11aff0a8e4f7af47bacce1dec27b0af
