+++ Assets/Scripts/Layers/LayerManager.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 
 using PAC.Animation;
-using PAC.Colour;
+using PAC.Colour.Compositing;
 using PAC.Extensions;
 using PAC.Files;
 using PAC.UI;
