+++ Assets/Tests/Colour/AlphaCompositingMode_Tests.cs
@@ -1,11 +1,12 @@
 using NUnit.Framework;
 
 using PAC.Colour;
+using PAC.Colour.Compositing;
 using PAC.Extensions;
 
 using UnityEngine;
 
-namespace PAC.Tests.Colour
+namespace PAC.Tests.Colour.Compositing
 {
     /// <summary>
     /// Tests for <see cref="AlphaCompositing"/>.
